refactor: any fate for simul with bug
