+++ config/config/event_info2.yml
@@ -1,19 +1,20 @@
 #事件优先选择文件，主要是选择能战斗的选项
 event:
-- 就是阮
-- 作弊
-- 重投
-- 奇物
-- 序列扑满
-- 重击
-- 对抗引力
-- 机械匣子
 - 盏烛火
-- 想要的
-- 钻石盒
-- 听一听
+- 和序列扑满玩
+- 信仰星神
+- 绞杀它们
+- 概率而战
+- 安保团队
+- 成年巨兽
+- 名流高级社交
+- 翻开牌
+- 见识见识真正
+- 周期性魔王
+- 头号矿工
+- 加入糖
+- 轻轻搅拌
+- 杀死它们
+- 周围的虫群
+- 正面突入
 - 购买1个奇物
-- 宇宙碎片
-- 绕路
-- 悄悄离开
-- 慌忙逃走
