+++ config/config/event_info3.yml
@@ -1,13 +1,7 @@
 #事件优先选择文件，主要是选择奇物与避战的选项
 event:
 - 就是阮
-- 空荡荡
-- 下班时间
-- 目标销售
-- 临时雇佣
-- 跨物种联谊
-- 回顾星际
-- 打败大人物
+- 悄悄偷懒
 - 作弊
 - 重投
 - 奇物
